Start rewrite to use API
